feat: Enhance MLP and Matrix classes with detailed documentation and new features
- Added a new binary file src.zip to the project.
- Updated MLPNXOR.java to improve documentation and commented out debugging code.
- Enhanced MSE.java with detailed comments and a private constructor to prevent instantiation.
- Expanded Matrix.java with comprehensive Javadoc comments for constructors, methods, and operations, including element-wise operations, matrix multiplication, and utility functions.
- Improved MLP.java by adding documentation for methods related to weight and bias management, forward propagation, and backpropagation.
- Updated IDifferentiableFunction.java and its implementations (Sigmoid and Step) with detailed descriptions of the activation functions and their derivatives.

+++ src/math/Matrix.java
@@ -7,8 +7,13 @@ import java.util.function.DoubleBinaryOperator;
 import java.util.function.DoubleUnaryOperator;
 
 /**
+ * A matrix class for mathematical operations on 2D arrays of doubles.
+ * Supports various operations including element-wise operations, matrix
+ * multiplication,
+ * transposition, and functional transformations.
+ *
  * @author hdaniel@ualg.pt
- * @author Tomás Machado
+ * @author André Martins, António Matoso, Tomás Machado
  * @version 202511052002
  */
 public class Matrix {
@@ -16,12 +21,24 @@ public class Matrix {
    private int rows;
    private int cols;
 
+   /**
+    * Constructs a matrix with the specified dimensions, initialized with zeros.
+    *
+    * @param rows the number of rows
+    * @param cols the number of columns
+    */
    public Matrix(int rows, int cols) {
       this.data = new double[rows][cols];
       this.rows = rows;
       this.cols = cols;
    }
 
+   /**
+    * Constructs a matrix from a 2D array of doubles.
+    * Creates a deep copy of the provided data.
+    *
+    * @param data the 2D array of values to initialize the matrix with
+    */
    public Matrix(double[][] data) {
       this.rows = data.length;
       this.cols = data[0].length;
@@ -32,6 +49,11 @@ public class Matrix {
       }
    }
 
+   /**
+    * Copy constructor. Creates a deep copy of another matrix.
+    *
+    * @param other the matrix to copy
+    */
    public Matrix(Matrix other) {
       this.rows = other.rows;
       this.cols = other.cols;
@@ -42,6 +64,15 @@ public class Matrix {
       }
    }
 
+   /**
+    * Creates a matrix with random values between 0 and 1.
+    *
+    * @param rows the number of rows
+    * @param cols the number of columns
+    * @param seed the random seed for reproducibility; if negative, uses current
+    *             time
+    * @return a new matrix filled with random values
+    */
    public static Matrix rand(int rows, int cols, int seed) {
       Matrix out = new Matrix(rows, cols);
       if (seed < 0) {
@@ -56,18 +87,40 @@ public class Matrix {
       return out;
    }
 
+   /**
+    * Gets the value at the specified position in the matrix.
+    *
+    * @param row the row index
+    * @param col the column index
+    * @return the value at the specified position
+    */
    public double get(int row, int col) {
       return this.data[row][col];
    }
 
+   /**
+    * Returns the number of rows in the matrix.
+    *
+    * @return the number of rows
+    */
    public int rows() {
       return this.rows;
    }
 
+   /**
+    * Returns the number of columns in the matrix.
+    *
+    * @return the number of columns
+    */
    public int cols() {
       return this.cols;
    }
 
+   /**
+    * Sets this matrix to be a copy of another matrix.
+    *
+    * @param other the matrix to copy from
+    */
    public void set(Matrix other) {
       this.data = other.data;
       this.rows = other.rows;
@@ -77,6 +130,12 @@ public class Matrix {
    // ==============================================================
    // Element operations
    // ==============================================================
+   /**
+    * Applies a function to each element of the matrix and returns a new matrix.
+    *
+    * @param fnc the function to apply to each element
+    * @return a new matrix with the function applied to all elements
+    */
    private Matrix traverse(DoubleUnaryOperator fnc) {
       Matrix result = new Matrix(rows, cols);
       for (int i = 0; i < rows; ++i) {
@@ -87,6 +146,11 @@ public class Matrix {
       return result;
    }
 
+   /**
+    * Applies a function to each element of this matrix in place.
+    *
+    * @param fnc the function to apply to each element
+    */
    private void traverseInPlace(DoubleUnaryOperator fnc) {
       for (int i = 0; i < rows; ++i) {
          for (int j = 0; j < cols; ++j) {
@@ -95,34 +159,78 @@ public class Matrix {
       }
    }
 
+   /**
+    * Applies a function to each element of the matrix and returns a new matrix.
+    *
+    * @param fnc the function to apply to each element
+    * @return a new matrix with the function applied to all elements
+    */
    public Matrix apply(DoubleUnaryOperator fnc) {
       return this.traverse(fnc);
    }
 
+   /**
+    * Applies a function to each element of this matrix in place.
+    *
+    * @param fnc the function to apply to each element
+    */
    public void applyInPlace(DoubleUnaryOperator fnc) {
       this.traverseInPlace(fnc);
    }
 
+   /**
+    * Multiplies each element of the matrix by a scalar and returns a new matrix.
+    *
+    * @param scalar the scalar value to multiply by
+    * @return a new matrix with all elements multiplied by the scalar
+    */
    public Matrix mult(double scalar) {
       return this.traverse(x -> x * scalar);
    }
 
+   /**
+    * Multiplies each element of this matrix by a scalar in place.
+    *
+    * @param scalar the scalar value to multiply by
+    */
    public void multInPlace(double scalar) {
       this.traverseInPlace(x -> x * scalar);
    }
 
+   /**
+    * Adds a scalar to each element of the matrix and returns a new matrix.
+    *
+    * @param scalar the scalar value to add
+    * @return a new matrix with the scalar added to all elements
+    */
    public Matrix add(double scalar) {
       return this.traverse(x -> x + scalar);
    }
 
+   /**
+    * Adds a scalar to each element of this matrix in place.
+    *
+    * @param scalar the scalar value to add
+    */
    public void addInPlace(double scalar) {
       this.traverseInPlace(x -> x + scalar);
    }
 
+   /**
+    * Subtracts a scalar from each element of the matrix and returns a new matrix.
+    *
+    * @param scalar the scalar value to subtract
+    * @return a new matrix with the scalar subtracted from all elements
+    */
    public Matrix sub(double scalar) {
       return this.traverse(x -> x - scalar);
    }
 
+   /**
+    * Subtracts a scalar from each element of this matrix in place.
+    *
+    * @param scalar the scalar value to subtract
+    */
    public void subInPlace(double scalar) {
       this.traverseInPlace(x -> x - scalar);
    }
@@ -130,6 +238,15 @@ public class Matrix {
    // ==============================================================
    // Element-wise operations between two matrices
    // ==============================================================
+   /**
+    * Applies a binary function element-wise between this matrix and another
+    * matrix.
+    *
+    * @param other the other matrix
+    * @param fnc   the binary function to apply
+    * @return a new matrix with the function applied element-wise
+    * @throws IllegalArgumentException if matrix dimensions don't match
+    */
    private Matrix elementWise(Matrix other, DoubleBinaryOperator fnc) {
       if (this.rows != other.rows || this.cols != other.cols) {
          throw new IllegalArgumentException(
@@ -144,6 +261,14 @@ public class Matrix {
       return result;
    }
 
+   /**
+    * Applies a binary function element-wise between this matrix and another matrix
+    * in place.
+    *
+    * @param other the other matrix
+    * @param fnc   the binary function to apply
+    * @throws IllegalArgumentException if matrix dimensions don't match
+    */
    private void elementWiseInPlace(Matrix other,
          DoubleBinaryOperator fnc) {
       if (this.rows != other.rows || this.cols != other.cols) {
@@ -157,26 +282,67 @@ public class Matrix {
       }
    }
 
+   /**
+    * Adds another matrix element-wise and returns a new matrix.
+    *
+    * @param other the matrix to add
+    * @return a new matrix with element-wise sum
+    * @throws IllegalArgumentException if matrix dimensions don't match
+    */
    public Matrix add(Matrix other) {
       return this.elementWise(other, (a, b) -> a + b);
    }
 
+   /**
+    * Adds another matrix element-wise to this matrix in place.
+    *
+    * @param other the matrix to add
+    * @throws IllegalArgumentException if matrix dimensions don't match
+    */
    public void addInPlace(Matrix other) {
       this.elementWiseInPlace(other, (a, b) -> a + b);
    }
 
+   /**
+    * Multiplies another matrix element-wise (Hadamard product) and returns a new
+    * matrix.
+    *
+    * @param other the matrix to multiply
+    * @return a new matrix with element-wise product
+    * @throws IllegalArgumentException if matrix dimensions don't match
+    */
    public Matrix mult(Matrix other) {
       return this.elementWise(other, (a, b) -> a * b);
    }
 
+   /**
+    * Multiplies another matrix element-wise (Hadamard product) to this matrix in
+    * place.
+    *
+    * @param other the matrix to multiply
+    * @throws IllegalArgumentException if matrix dimensions don't match
+    */
    public void multInPlace(Matrix other) {
       this.elementWiseInPlace(other, (a, b) -> a * b);
    }
 
+   /**
+    * Subtracts another matrix element-wise and returns a new matrix.
+    *
+    * @param other the matrix to subtract
+    * @return a new matrix with element-wise difference
+    * @throws IllegalArgumentException if matrix dimensions don't match
+    */
    public Matrix sub(Matrix other) {
       return this.elementWise(other, (a, b) -> a - b);
    }
 
+   /**
+    * Subtracts another matrix element-wise from this matrix in place.
+    *
+    * @param other the matrix to subtract
+    * @throws IllegalArgumentException if matrix dimensions don't match
+    */
    public void subInPlace(Matrix other) {
       this.elementWiseInPlace(other, (a, b) -> a - b);
    }
@@ -184,6 +350,11 @@ public class Matrix {
    // ==============================================================
    // Other math operations
    // ==============================================================
+   /**
+    * Computes the sum of all elements in the matrix.
+    *
+    * @return the sum of all matrix elements
+    */
    public double sum() {
       double total = 0.0;
       for (int i = 0; i < this.rows; ++i) {
@@ -194,6 +365,11 @@ public class Matrix {
       return total;
    }
 
+   /**
+    * Computes the sum of each column and returns a row vector (1 x cols matrix).
+    *
+    * @return a row vector containing the sum of each column
+    */
    public Matrix sumColumns() {
       Matrix result = new Matrix(1, this.cols);
       for (int j = 0; j < this.cols; ++j) {
@@ -206,6 +382,15 @@ public class Matrix {
       return result;
    }
 
+   /**
+    * Performs matrix multiplication (dot product) with another matrix.
+    *
+    * @param other the matrix to multiply with
+    * @return a new matrix representing the matrix product
+    * @throws IllegalArgumentException if the number of columns in this matrix
+    *                                  doesn't match the number of rows in the
+    *                                  other matrix
+    */
    public Matrix dot(Matrix other) {
       if (this.cols != other.rows) {
          throw new IllegalArgumentException(
@@ -222,6 +407,15 @@ public class Matrix {
       return result;
    }
 
+   /**
+    * Adds a row vector to each row of the matrix and returns a new matrix.
+    * This is useful for adding biases in neural networks.
+    *
+    * @param rowVector a 1 x cols matrix to add to each row
+    * @return a new matrix with the row vector added to each row
+    * @throws IllegalArgumentException if rowVector is not a row vector or
+    *                                  doesn't have the same number of columns
+    */
    public Matrix addRowVector(Matrix rowVector) {
       if (rowVector.rows() != 1 || rowVector.cols() != this.cols) {
          throw new IllegalArgumentException(
@@ -236,6 +430,14 @@ public class Matrix {
       return result;
    }
 
+   /**
+    * Adds a row vector to each row of this matrix in place.
+    * This is useful for adding biases in neural networks.
+    *
+    * @param rowVector a 1 x cols matrix to add to each row
+    * @throws IllegalArgumentException if rowVector is not a row vector or
+    *                                  doesn't have the same number of columns
+    */
    public void addInPlaceRowVector(Matrix rowVector) {
       if (rowVector.rows() != 1 || rowVector.cols() != this.cols) {
          throw new IllegalArgumentException(
@@ -251,6 +453,12 @@ public class Matrix {
    // ==============================================================
    // Column and row operations
    // ==============================================================
+   /**
+    * Returns the transpose of this matrix.
+    * Rows and columns are swapped.
+    *
+    * @return a new matrix that is the transpose of this matrix
+    */
    public Matrix transpose() {
       Matrix result = new Matrix(cols, rows);
       // transpose the matrix
@@ -266,6 +474,14 @@ public class Matrix {
    // ==============================================================
    // Compare operations
    // ==============================================================
+   /**
+    * Compares this matrix with another object for equality.
+    * Two matrices are equal if they have the same dimensions and all elements are
+    * equal.
+    *
+    * @param o the object to compare with
+    * @return true if the matrices are equal, false otherwise
+    */
    @Override
    public boolean equals(Object o) {
       if (!(o instanceof Matrix matrix)) {
@@ -275,6 +491,11 @@ public class Matrix {
             Objects.deepEquals(this.data, matrix.data);
    }
 
+   /**
+    * Returns a hash code value for this matrix.
+    *
+    * @return a hash code based on the matrix data, rows, and columns
+    */
    @Override
    public int hashCode() {
       return Objects.hash(Arrays.deepHashCode(this.data), this.rows,
@@ -284,6 +505,13 @@ public class Matrix {
    // ==============================================================
    // Convert operations
    // ==============================================================
+   /**
+    * Converts the matrix to a string representation with elements cast to
+    * integers.
+    * Each row is on a separate line with values separated by spaces.
+    *
+    * @return a string representation with integer values
+    */
    public String toIntString() {
       StringBuilder sb = new StringBuilder();
       for (int i = 0; i < this.rows; ++i) {
@@ -296,11 +524,17 @@ public class Matrix {
       return sb.toString();
    }
 
+   /**
+    * Returns a string representation of the matrix.
+    * Each row is on a separate line with values separated by spaces.
+    *
+    * @return a string representation of the matrix
+    */
    @Override
    public String toString() {
       StringBuilder sb = new StringBuilder();
       for (int i = 0; i < this.rows; ++i) {
-         sb.append((int) this.data[i][0]);
+         sb.append(this.data[i][0]);
          for (int j = 1; j < this.cols; ++j) {
             sb.append(" ").append(this.data[i][j]);
          }
